feat: allow zero delta mouse events to update simulation
As of
https://github.com/kovidgoyal/kitty/commit/27df509206675ba49480d646a4926e5a74990eb3
kitty now always sends an CSI mouse event when the terminal regains
focus and is configured to track motion. Accepting zero delta mouse
events allows for kitty as a background panel with awrit and webgl-fluid
to visual indicate when moving to a workspace with no surfaces even if
the mouse hasn't moved since the terminal last lost focus. (pretty poofs
when navigating from a surface to the background panel)